requirements.txt
Flask==3.0.3
reportlab==4.2.5
gunicorn==23.0.0
python-dotenv==1.0.1
PyMySQL==1.1.1
